# nest-backend/openApiSpec.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/access-control/users/me/rules": {
            "get": {
                "operationId": "AccessControlController_getCurrentUserRules",
                "parameters": [
                    {
                        "name": "feature",
                        "required": false,
                        "in": "query",
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "name",
                        "required": false,
                        "in": "query",
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "subject",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "enum": [
                                "Poster",
                                "Role",
                                "User",
                                "Tenant",
                                "Workspace",
                                "Project",
                                "Comment",
                                "Invite",
                                "Task",
                                "Board"
                            ],
                            "type": "string"
                        }
                    },
                    {
                        "name": "action",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "enum": [
                                    "create",
                                    "read",
                                    "list",
                                    "update",
                                    "delete",
                                    "manage",
                                    "joinBeforeStartData",
                                    "requestCollaboration",
                                    "join"
                                ]
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": { "type": "object" }
                                }
                            }
                        }
                    }
                },
                "tags": ["Access Control"]
            }
        },
        "/access-control/users/{id}/rules": {
            "get": {
                "operationId": "AccessControlController_getUserRules",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    },
                    {
                        "name": "feature",
                        "required": false,
                        "in": "query",
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "name",
                        "required": false,
                        "in": "query",
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "subject",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "enum": [
                                "Poster",
                                "Role",
                                "User",
                                "Tenant",
                                "Workspace",
                                "Project",
                                "Comment",
                                "Invite",
                                "Task",
                                "Board"
                            ],
                            "type": "string"
                        }
                    },
                    {
                        "name": "action",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "enum": [
                                    "create",
                                    "read",
                                    "list",
                                    "update",
                                    "delete",
                                    "manage",
                                    "joinBeforeStartData",
                                    "requestCollaboration",
                                    "join"
                                ]
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": { "type": "object" }
                                }
                            }
                        }
                    }
                },
                "tags": ["Access Control"]
            },
            "post": {
                "operationId": "AccessControlController_createRuleForUser",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRuleDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": { "type": "object" }
                            }
                        }
                    }
                },
                "tags": ["Access Control"]
            }
        },
...
        "/workspaces/{workspaceId}/posters": {
            "get": {
                "operationId": "Poster_find_many",
                "summary": "",
                "parameters": [
                    {
                        "name": "workspaceId",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    },
                    {
                        "name": "title",
                        "required": false,
                        "in": "query",
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "topics",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "array",
                            "items": { "type": "string" }
                        }
                    },
                    {
                        "name": "keywords",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "array",
                            "items": { "type": "string" }
                        }
                    },
                    {
                        "name": "authors",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "array",
                            "items": { "type": "string" }
                        }
                    },
                    {
                        "name": "organizations",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "array",
                            "items": { "type": "string" }
                        }
                    },
                    {
                        "name": "search",
                        "required": true,
                        "in": "query",
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "or",
                        "required": false,
                        "in": "query",
                        "schema": { "default": false, "type": "boolean" }
                    },
                    {
                        "name": "take",
                        "required": false,
                        "in": "query",
                        "schema": { "type": "number" }
                    },
                    {
                        "name": "skip",
                        "required": false,
                        "in": "query",
                        "schema": { "type": "number" }
                    }
                ],
                "responses": { "200": { "description": "" } },
                "tags": ["Posters"]
            },
            "post": {
                "operationId": "Poster_create",
                "summary": "",
                "parameters": [
                    {
                        "name": "workspaceId",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePosterDto"
                            }
                        }
                    }
                },
                "responses": { "201": { "description": "" } },
                "tags": ["Posters"]
            }
        },
        "/workspaces/{workspaceId}/posters/{id}": {
            "get": {
                "operationId": "Poster_find_one",
                "summary": "",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    },
                    {
                        "name": "workspaceId",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    }
                ],
                "responses": { "200": { "description": "" } },
                "tags": ["Posters"]
            },
            "patch": {
                "operationId": "Poster_update",
                "summary": "",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    },
                    {
                        "name": "workspaceId",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePosterDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Poster"
                                }
                            }
                        }
                    }
                },
                "tags": ["Posters"]
            },
            "delete": {
                "operationId": "Poster_delete",
                "summary": "",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    },
                    {
                        "name": "workspaceId",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    }
                ],
                "responses": { "200": { "description": "" } },
                "tags": ["Posters"]
            }
        },
...
    },
    "info": {
        "title": "Posters example",
        "description": "The posters API description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
...
            "Upload": {
                "type": "object",
                "properties": {
                    "id": { "type": "number" },
                    "fieldname": { "type": "string" },
                    "originalname": { "type": "string" },
                    "encoding": { "type": "string" },
                    "contentEncoding": { "type": "string" },
                    "mimetype": { "type": "string" },
                    "size": { "type": "number" },
                    "bucket": { "type": "string" },
                    "key": { "type": "string" },
                    "acl": { "type": "string" },
                    "contentType": { "type": "string" },
                    "contentDisposition": { "type": "string" },
                    "storageClass": { "type": "string" },
                    "serverSideEncryption": { "type": "string" },
                    "metadata": { "type": "object" },
                    "location": { "type": "string" },
                    "etag": { "type": "string" },
                    "versionId": { "type": "string" }
                },
                "required": [
                    "id",
                    "fieldname",
                    "originalname",
                    "encoding",
                    "contentEncoding",
                    "mimetype",
                    "size",
                    "bucket",
                    "key",
                    "acl",
                    "contentType",
                    "contentDisposition",
                    "storageClass",
                    "serverSideEncryption",
                    "metadata",
                    "location",
                    "etag",
                    "versionId"
                ]
            },
...
            "CreateRuleDto": {
                "type": "object",
                "properties": {
                    "feature": { "type": "string" },
                    "name": { "type": "string" },
                    "options": { "type": "object" },
                    "userId": { "type": "number" }
                },
                "required": ["feature", "name", "options"]
            },
...
            "CreatePosterDataWorkspace": {
                "type": "object",
                "properties": {
                    "id": { "type": "number" },
                    "tenantId": { "type": "number" }
                },
                "required": ["id", "tenantId"]
            },
            "KeywordDto": {
                "type": "object",
                "properties": { "name": { "type": "string" } },
                "required": ["name"]
            },
            "TopicDto": {
                "type": "object",
                "properties": { "name": { "type": "string" } },
                "required": ["name"]
            },
            "AuthorDto": {
                "type": "object",
                "properties": { "id": { "type": "number" } },
                "required": ["id"]
            },
            "CreatePosterData": {
                "type": "object",
                "properties": {
                    "title": { "type": "string" },
                    "workspace": {
                        "$ref": "#/components/schemas/CreatePosterDataWorkspace"
                    },
                    "description": { "type": "string" },
                    "keywords": {
                        "type": "array",
                        "items": { "$ref": "#/components/schemas/KeywordDto" }
                    },
                    "topics": {
                        "type": "array",
                        "items": { "$ref": "#/components/schemas/TopicDto" }
                    },
                    "visibility": { "type": "object" },
                    "authors": {
                        "type": "array",
                        "items": { "$ref": "#/components/schemas/AuthorDto" }
                    }
                },
                "required": ["title", "workspace", "visibility", "authors"]
            },
            "CreatePosterDto": {
                "type": "object",
                "properties": {
                    "image": {
                        "type": "string",
                        "format": "binary",
                        "description": ".jpeg file"
                    },
                    "pdf": {
                        "type": "string",
                        "format": "binary",
                        "description": ".pdf file"
                    },
                    "data": {
                        "description": "Schema: CreatePosterData",
                        "allOf": [
                            { "$ref": "#/components/schemas/CreatePosterData" }
                        ]
                    }
                },
                "required": ["image", "pdf", "data"]
            },
            "PosterUpdateUploadDto": {
                "type": "object",
                "properties": { "id": { "type": "number" } },
                "required": ["id"]
            },
            "UpdatePosterData": {
                "type": "object",
                "properties": {
                    "pdf": {
                        "$ref": "#/components/schemas/PosterUpdateUploadDto"
                    },
                    "image": {
                        "$ref": "#/components/schemas/PosterUpdateUploadDto"
                    },
                    "title": { "type": "string" },
                    "workspace": {
                        "$ref": "#/components/schemas/CreatePosterDataWorkspace"
                    },
                    "description": { "type": "string" },
                    "keywords": {
                        "type": "array",
                        "items": { "$ref": "#/components/schemas/KeywordDto" }
                    },
                    "topics": {
                        "type": "array",
                        "items": { "$ref": "#/components/schemas/TopicDto" }
                    },
                    "visibility": { "type": "object" },
                    "authors": {
                        "type": "array",
                        "items": { "$ref": "#/components/schemas/AuthorDto" }
                    }
                }
            },
            "UpdatePosterDto": {
                "type": "object",
                "properties": {
                    "image": {
                        "type": "string",
                        "format": "binary",
                        "description": ".jpeg file"
                    },
                    "pdf": {
                        "type": "string",
                        "format": "binary",
                        "description": ".pdf file"
                    },
                    "data": {
                        "description": "Schema: UpdatePosterData",
                        "allOf": [
                            { "$ref": "#/components/schemas/UpdatePosterData" }
                        ]
                    }
                },
                "required": ["image", "pdf", "data"]
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": { "type": "number" },
                    "objectId": { "type": "string" },
                    "organizationName": { "type": "string", "nullable": true },
                    "firstName": { "type": "string" },
                    "lastName": { "type": "string" },
                    "email": { "type": "string" },
                    "isDummyUser": { "type": "boolean" },
                    "subscribed": { "type": "boolean" },
                    "privacyPolicy": { "type": "boolean" },
                    "termOfUse": { "type": "boolean" },
                    "disableNotifications": { "type": "boolean" },
                    "linkedin": { "type": "string" },
                    "researchGate": { "type": "string" },
                    "twitter": { "type": "string" },
                    "currentPosition": { "type": "string", "nullable": true },
                    "biography": { "type": "string", "nullable": true },
                    "createdAt": { "format": "date-time", "type": "string" },
                    "updatedAt": { "format": "date-time", "type": "string" },
                    "avatarId": { "type": "number", "nullable": true },
                    "avatar": { "$ref": "#/components/schemas/Upload" },
                    "posters": {
                        "type": "array",
                        "items": { "type": "object" }
                    },
                    "authoredPosters": {
                        "type": "array",
                        "items": { "type": "object" }
                    },
                    "roles": { "type": "array", "items": { "type": "object" } },
                    "comments": {
                        "type": "array",
                        "items": { "type": "object" }
                    },
                    "userDirectory": { "type": "object" },
                    "skills": { "type": "array", "items": { "type": "object" } }
                },
                "required": [
                    "id",
                    "objectId",
                    "organizationName",
                    "firstName",
                    "lastName",
                    "email",
                    "isDummyUser",
                    "subscribed",
                    "privacyPolicy",
                    "termOfUse",
                    "disableNotifications",
                    "linkedin",
                    "researchGate",
                    "twitter",
                    "currentPosition",
                    "biography",
                    "createdAt",
                    "updatedAt",
                    "avatarId"
                ]
            },
            "Keyword": {
                "type": "object",
                "properties": { "name": { "type": "string" } },
                "required": ["name"]
            },
            "Topic": {
                "type": "object",
                "properties": { "name": { "type": "string" } },
                "required": ["name"]
            },
            "Poster": {
                "type": "object",
                "properties": {
                    "id": { "type": "number" },
                    "user": { "$ref": "#/components/schemas/User" },
                    "userId": { "type": "number" },
                    "workspaceId": { "type": "number" },
                    "comments": {
                        "type": "array",
                        "items": { "type": "object" }
                    },
                    "title": { "type": "string" },
                    "description": { "type": "string" },
                    "pdf": { "type": "object" },
                    "pdfId": { "type": "number" },
                    "image": { "type": "object" },
                    "imageId": { "type": "number" },
                    "visibility": { "type": "object" },
                    "keywords": {
                        "type": "array",
                        "items": { "$ref": "#/components/schemas/Keyword" }
                    },
                    "topics": {
                        "type": "array",
                        "items": { "$ref": "#/components/schemas/Topic" }
                    },
                    "authors": {
                        "type": "array",
                        "items": { "type": "object" }
                    },
                    "createdAt": { "format": "date-time", "type": "string" },
                    "updatedAt": { "format": "date-time", "type": "string" }
                },
                "required": [
                    "id",
                    "userId",
                    "workspaceId",
                    "title",
                    "description",
                    "pdf",
                    "pdfId",
                    "image",
                    "imageId",
                    "visibility",
                    "keywords",
                    "topics",
                    "authors",
                    "createdAt",
                    "updatedAt"
                ]
            },
...
        }
    }
}